requirements.txt
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
PyQt5>=5.15.9
requests>=2.31.0
python-dateutil>=2.8.2
